+++ src/game/creature.h
@@ -143,7 +143,7 @@ public:
 
 	virtual void updateSprite() = 0;
 
-	void forceMove( Position& to );
+	void forceMove( const Position& to );
 
 	virtual CreatureTickResult onTick( quint64 tickNumber, bool seasonChanged, bool dayChanged, bool hourChanged, bool minuteChanged ) = 0;